fix delete old strm

+++ clases/folders/folders.py
@@ -14,38 +14,65 @@ class folders:
     if 'ytdlp2strm_temp_file_duration' in ytdlp2strm_config:
         keep_downloaded = int(ytdlp2strm_config['ytdlp2strm_temp_file_duration'])
 
+class folders:
+    ytdlp2strm_config = c.config('./config/config.json').get_config()
+
+    keep_downloaded = 86400
+    temp_aria2_ffmpeg_files = 600
+    if 'ytdlp2strm_temp_file_duration' in ytdlp2strm_config:
+        keep_downloaded = int(ytdlp2strm_config['ytdlp2strm_temp_file_duration'])
+
     def make_clean_folder(self, folder_path, forceclean, config):
         if os.path.exists(folder_path):
-            if forceclean or not config.get("ytdlp2strm_keep_old_strm", True):
-                file_list = glob.glob(os.path.join(folder_path, "*"))
-                for file_path in file_list:
+            if forceclean or not config.get("ytdlp2strm_keep_old_strm"):
+                # Check the contents of the directory in a simpler way
+                try:
+                    file_list = os.listdir(folder_path)
+                except Exception as e:
+                    return
+
+                now = time.time()
+
+                for file_name in file_list:
+                    file_path = os.path.join(folder_path, file_name)
                     if os.path.isfile(file_path):
-                        os.remove(file_path)
-                        log_text = (f"Deleted file: {file_path}")
-                        l.log("folder", log_text)
-                log_text = (f"Cleaned directory: {folder_path}")
+                        file_age = now - self.modified_date(file_path)
+
+                        if forceclean or file_age > self.keep_downloaded:
+                            try:
+                                os.remove(file_path)
+                                log_text = f"Deleted file: {file_path}"
+                                l.log("folder", log_text)
+                                print(log_text)
+                            except Exception as e:
+                                log_text = f"Failed to delete file: {file_path}. Error: {e}"
+                                l.log("folder", log_text)
+                log_text = f"Cleaned directory: {folder_path}"
                 l.log("folder", log_text)
         else:
+            print(f"Folder does not exist, creating: {folder_path}")
             os.makedirs(folder_path, exist_ok=True)
-            log_text = (f"Created directory: {folder_path}")
+            log_text = f"Created directory: {folder_path}"
             l.log("folder", log_text)
+            print(log_text)
+
 
     def write_file(self, file_path, content):
         try:
-            # Ensure content is properly encoded
-            content = content.encode('utf-8').decode('utf-8')
-            
-            # Write to file with UTF-8 encoding
-            with open(file_path, "w", encoding="utf-8") as file:
-                file.write(content.replace('\n',''))
-            
-            file_path = file_path.encode('utf-8').decode('utf-8')
-            log_text = f"File created: {file_path}"
-            l.log("folder", log_text)
+            if not os.path.exists(file_path) or 'tvshow.nfo' in file_path:
+                # Ensure content is properly encoded
+                content = content.encode('utf-8').decode('utf-8')
+                
+                # Write to file with UTF-8 encoding
+                with open(file_path, "w", encoding="utf-8") as file:
+                    file.write(content.replace('\n',''))
+                
+                file_path = file_path.encode('utf-8').decode('utf-8')
+                log_text = f"File created: {file_path}"
+                l.log("folder", log_text)
         except Exception as e:
             log_text = f"Error writing file: {e}"
             l.log("folder", log_text)
-            pass
 
     def clean_waste(self, files_to_delete):
         for file_path in files_to_delete:
@@ -59,21 +86,22 @@ class folders:
                 l.log("folder", log_text)
                 pass
 
-    def clean_old_videos(self, stop_event):
-        def creation_date(path_to_file):
-            if platform.system() == 'Windows':
-                return os.path.getctime(path_to_file)
-            else:
-                stat = os.stat(path_to_file)
-                try:
-                    return stat.st_birthtime
-                except AttributeError:
-                    return stat.st_mtime
-
-        def modified_date(path_to_file):
+    def creation_date(self, path_to_file):
+        if platform.system() == 'Windows':
+            return os.path.getctime(path_to_file)
+        else:
             stat = os.stat(path_to_file)
-            return stat.st_mtime
-        
+            try:
+                return stat.st_birthtime
+            except AttributeError:
+                return stat.st_mtime
+
+    def modified_date(self, path_to_file):
+        stat = os.stat(path_to_file)
+        return stat.st_mtime
+    
+    def clean_old_videos(self, stop_event):
+
         while not stop_event.is_set():
             try:
                 time.sleep(5)
@@ -86,12 +114,12 @@ class folders:
                     temp_file = os.path.join(temp_path, f)
                     if not f == "__init__.py":
                         if any(keyword in f for keyword in aria2_ffmpeg_files):
-                            if os.path.isfile(temp_file) and modified_date(temp_file) < now - self.temp_aria2_ffmpeg_files:
+                            if os.path.isfile(temp_file) and self.modified_date(temp_file) < now - self.temp_aria2_ffmpeg_files:
                                 log_text = (f"Removing old temporary file: {temp_file}")
                                 l.log("folder", log_text)
                                 os.remove(temp_file)
                         else:
-                            if os.path.isfile(temp_file) and modified_date(temp_file) < now - self.keep_downloaded:
+                            if os.path.isfile(temp_file) and self.modified_date(temp_file) < now - self.keep_downloaded:
                                 log_text = (f"Removing old video file: {temp_file}")
                                 l.log("folder", log_text)
                                 os.remove(temp_file)
